// src/app/state.enum.ts
export enum State {
    WaitForScan = 'waitScan',
    WaitForCode = 'waitForCode',
    FindProduct = 'findProduct',
    SelectProduct = 'selectProduct',
    ErrorUnknowPdt = 'errorUnknowPdt',
    SelectAmount = 'selectAmount',
    EditProduct = 'editProduct',
    ChoosePayMode = 'choosePayMode',
    AmountToPay = 'amountToPay',
    CashAmount = 'cashAmount',
    CashOut = 'cashOut',
    PutOnHold = 'putOnHold',
}